refactor: split gateway operator read models
